Fix the parsing for the leagues that are in play offs.
During play off, the team have the final standings attached
to the end of the team name. ex: Kelly(9). The fix is to extract
the name before '('.

Also, before adding the team to the database, make sure it's not in
the table already.

Special names (PLAYOFF) are no longer will be stored in the teams
database.

Signed-off-by: Boris Dubinsky <dubi13@gmail.com>